requirements.txt
 llama-index-llms-ollama
 llama-index-embeddings-ollama
 llama-index-vector-stores-faiss
faiss-cpu
